requirements.txt
anthropic>=0.49.0
python-docx>=1.1.0
gradio>=4.44.0
lxml>=5.2.0
pydantic>=2.7.0
python-dateutil>=2.9.0
openpyxl>=3.1.0
pypdf>=4.0.0
pdfplumber>=0.11.0
